requirements.txt
Django>=4.2
djangorestframework>=3.14
python-dotenv>=1.0.1
openai>=1.30.1
django-cors-headers>=4.3.1
